Review fitting interface and parameter sync
The fitting interface is not more general and allows to decide the minimizer to be used in minuit. The old Migrad interface has been kept for backward compatibility but it will be removed in future versions
Changed deafult in the parameter syncronization. Now the starting value of the fit is always taken from the default value. Optionally it can be left defined by the last minimization